requirements.txt
openai>=1.12.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
ttkbootstrap>=1.10.1
